Revise split time and add to protocol detail pages
